+++ app/models/renalware/diaverum/incoming/session_xml_document.rb
@@ -10,6 +10,15 @@ module Renalware
         pattr_initialize :node
         delegate :xpath, to: :node
 
+        # Most node attributes and handled in method_missing but StartTime is
+        # a special case because we want to stub in tests and it therefore needs to be
+        # on the object's interface.
+        # rubocop:disable Naming/MethodName
+        def StartTime
+          node.xpath("StartTime")&.text
+        end
+        # rubocop:enable Naming/MethodName
+
         # Delegate methods names in CamelCase to the decorated XML Treatment node
         # so we can use e.g. session_xml_doc.TreatmentId
         def method_missing(method, *args, &block)
