+++ config/initializers/inflections.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+ActiveSupport::Inflector.inflections(:en) do |inflect|
+  inflect.acronym "DNA"
+end
